refactor: DinamicForm + minor changes
